requirements.txt
aio-pika==9.5.5
httpx~=0.28.1
Markdown~=3.7.0
mistralai~=1.3.1
mistral_common~=1.5.4
numpy==2.2.4
openpyxl>=3.1.5
pandas==2.2.3
plotly==6.0.0
pika==1.3.2
psycopg2-binary~=2.9.10
pydantic>=2.7.1,<3.0.0
python_docx~=1.1.2
python-dotenv~=1.0.1
python-telegram-bot~=21.10
python_dateutil==2.8.2
PyYAML~=6.0.2
requests~=2.32.1
scikit-image==0.25.2
streamlit~=1.41.0
SQLAlchemy==2.0.37
git+https://github.com/Elpharran/docling.git
